Use getNick() in Minebot instead of this.nickname
The former is updated on every nick change, including those done by the
server.

This fixes a bug where the bot has nick "nick" defined, but it was
already used on login and the bot switches to "nick2". On every
PART/QUIT/KICK event by "nick" was interpreted as craftirc leaving the
channel.

Also changed delivery() to check if an EndPoint is null before adding it
to the destination list (This caused a null pointer exception before the
main bug was fixed)